requirements.txt
# Core Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
# Use psycopg (modern PostgreSQL adapter) instead of psycopg2-binary for better Windows compatibility
psycopg[binary]==3.1.12
dj-database-url==2.1.0

# Alternative database options (uncomment if PostgreSQL installation issues persist)
# For SQLite (development): No additional package needed (built into Python)
# For MySQL: mysqlclient==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Image Processing (for portfolio images)
Pillow==10.1.0

# Email Backend
django-anymail==10.2

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Environment Management
python-decouple==3.8

# Utilities
requests==2.31.0
python-slugify==8.0.1